+++ api/index.js
@@ -2,11 +2,11 @@ const server = require("./src/app.js");
 const { conn } = require("./src/db.js");
 const {
   fillSupervisorShifts,
-} = require("./src/Controllers/getSupervisorShift.js");
+} = require("./src/Controllers/Supervisor/getSupervisorShift.js");
 const {
   fillCompanionShifts,
-} = require("./src/Controllers/getCompanionShift.js");
-const postCityTimeZone = require("./src/Controllers/postCityTimeZone.js");
+} = require("./src/Controllers/Companion/getCompanionShift.js");
+const postCityTimeZone = require("./src/Controllers/TimeZone/postCityTimeZone.js");
 // conn.sync({ alter: true }).then(async () => {
 conn.sync({ force: false }).then(async () => {
   await fillCompanionShifts();
