+++ api/src/Controllers/Supervisor/postSupervisor.js
@@ -1,4 +1,4 @@
-const { Supervisor } = require("../db");
+const { Supervisor } = require("../../db");
 const bcrypt = require("bcrypt");
 
 //Controlador para crear un Supervisor
