+++ api/src/Controllers/TimeZone/getCityTimeZone.js
@@ -1,4 +1,4 @@
-const { CityTimeZone } = require("../db");
+const { CityTimeZone } = require("../../db");
 
 //recupera todas las time zones de la base de datos y las devuelve un array de objetos del tipo:
 // {"id": "6a522905-8dc9-4bdc-84b1-af7f23514217", "zoneName": "Africa/Abidjan","offSet": "UTC+00:00"}
